airflow-batch-processor/scripts/requirements.txt
# JWT generation dependencies
PyJWT>=2.8.0
